write up dropout notebook
